Split v2 eval v1 bridge helpers
